+++ src/scenes/title_scene.lua
@@ -9,9 +9,18 @@ local startBtn = {
     x = 0,
     y = 0,
     width = 200,
-    height = 60,
+    height = 40,
     text = "START",
-    hovered = false
+    hovered = false,
+}
+
+local lboardBtn = {
+    x = 0,
+    y = 0,
+    width = 200,
+    height = 40,
+    text = "LEADERBOARD",
+    hovered = false,
 }
 
 function titleScene:load(assets, actions)
@@ -19,9 +28,15 @@ function titleScene:load(assets, actions)
     self.actions = actions
 end
 
+function titleScene:keypressed(key)
+    self.actions.quit()
+end
+
 function titleScene:mousepressed(x, y, btn)
     if btn == 1 and startBtn.hovered then
         self.actions.switchScene("main")
+    elseif btn == 1 and lboardBtn.hovered then
+        self.actions.switchScene("leaderboard")
     end
 end
 
@@ -30,17 +45,26 @@ function titleScene:update(dt)
     startBtn.hovered =
         mx > startBtn.x and mx < startBtn.x + startBtn.width and
         my > startBtn.y and my < startBtn.y + startBtn.height
+
+    lboardBtn.hovered =
+        mx > lboardBtn.x and mx < lboardBtn.x + lboardBtn.width and
+        my > lboardBtn.y and my < lboardBtn.y + lboardBtn.height
 end
 
 function titleScene:draw()
-    love.graphics.clear(colors)
+    love.graphics.clear(colors.BG)
 
     drawer:drawCenteredText("SLITHER", self.assets.mainFont, -28)
 
-    -- Draw centered button
+    -- Draw start button
     startBtn.x = (love.graphics.getWidth() - startBtn.width) / 2
-    startBtn.y = (love.graphics.getHeight() - startBtn.height) / 2 + 28
+    startBtn.y = (love.graphics.getHeight() - startBtn.height) / 2 + 48
     drawer:drawButton(startBtn, self.assets.subFont)
+
+    -- Draw leaderboard button
+    lboardBtn.x = startBtn.x
+    lboardBtn.y = startBtn.y + 48
+    drawer:drawButton(lboardBtn, self.assets.subFont)
 end
 
 return titleScene